Move brand_helper and add test file
