+++ lib/govuk_publishing_components.rb
@@ -1,6 +1,5 @@
 require "govuk_publishing_components/config"
 require "govuk_publishing_components/engine"
-require "govuk_publishing_components/presenters/brand_helper"
 require "govuk_publishing_components/presenters/contextual_navigation"
 require "govuk_publishing_components/presenters/related_navigation_helper"
 require "govuk_publishing_components/presenters/step_by_step_nav_helper"
@@ -15,6 +14,7 @@ require "govuk_publishing_components/presenters/curated_taxonomy_sidebar_links"
 require "govuk_publishing_components/presenters/content_item"
 
 require "govuk_publishing_components/app_helpers/taxon_breadcrumbs"
+require "govuk_publishing_components/app_helpers/brand_helper"
 
 module GovukPublishingComponents
 end
